Return Book and Register new User
Both these are now functional. Several modifications had to be made in
the backend to prevent crashing on null date values. Datatype of pending
fines has been changed to double.

+++ LMS/src/Client.java
@@ -6,7 +6,7 @@ import java.sql.Date;
 public class Client extends User 
 {
 
-	private int pendingFines;
+	private double pendingFines;
 	private String memberType;
 	private Date expiryDate;
 
@@ -18,7 +18,7 @@ public class Client extends User
 
 	}
 
-	public int getPendingFines(){
+	public double getPendingFines(){
 		return this.pendingFines;
 	}
 
@@ -30,7 +30,7 @@ public class Client extends User
 		return this.expiryDate;
 	}
 
-	public void setPendingFines(int pendingFines){
+	public void setPendingFines(double pendingFines){
 		this.pendingFines = pendingFines;
 	}
